move utilities for storage into utils directory
